+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu
 
 * dev-master
     * HOTFIX      #1962 [ContactBundle]     Fixed contact birthday trigger save-button
+    * HOTFIX      #1958 [SnippetBundle]     Fixed bug with snippet in snippet
 
 * 1.1.7 (2016-01-26)
     * HOTFIX      #1952 [PersistenceBundle] Fixed mapped superclass inheritance